requirements.txt
django==3.0.7
django-debug-toolbar==2.2
dj-database-url==0.5.0
Pillow==7.1.2
djangorestframework==3.12.*
python-dotenv==0.15.0
requests==2.24.0
environs==9.2.0
geopy==2.0.0
django-phonenumber-field=5.0.0
phonenumbers==8.12.20
